Relocating COMP20008 Assignment 1 to personal repository
This assignment until now has been sitting in a private repository that was generated while I was undertaking COMP20008 - Elements of Data Processing. I have retrieved the version where I literally added in an assignment to the current_path variable in partb1.py which the code completely breaks without. This unfortunately led to me losing all the marks for part B as I somehow overlooked the fact that I used an unassigned variable before submitting.